Image utils added.
All the tranformations applied for selectivity index have been
moved to image utils. Also ImageDataset new class has been
created for save some information about the dataset used by
the network.
