+++ nefesi/color_index.py
@@ -1,70 +1,13 @@
 import os
 import pickle
 
-import numpy as np
-from keras.preprocessing import image
-
 import read_activations
 from neuron_feature import get_image_receptive_field
+from util.image import rgb2opp, image2max_gray
 
 
 
-def load_images(dataset_path, image_names):
-    images = []
-    for n in image_names:
-        i = image.load_img(dataset_path + n, target_size=(224, 224))
-        i = image.img_to_array(i)
-        # i -= avg_img
-        images.append(i)
-
-    # i = image.array_to_img(images[0], scale=False)
-    # i.save('origin.png')
-    return images
-
-def rgb2opp(img):
-    """Converts an image from RGB space to OPP (Opponent color space).
-
-    :param img: Numpy array
-    :return: Numpy array
-    """
-
-    # x = image.img_to_array(img)
-    opp = np.zeros(shape=img.shape)
-
-    x = img / 255.
-    R = x[:, :, 0]
-    G = x[:, :, 1]
-    B = x[:, :, 2]
-    opp[:, :, 0] = (R + G + B - 1.5) / 1.5
-    opp[:, :, 1] = (R - G)
-    opp[:, :, 2] = (R + G - 2 * B) / 2
-
-    return opp
-
-def image2max_gray(img):
-    """Converts an image to gray scale using the PCA in order to maximally
-     preserve the shape pattern of the image.
-
-    :param img: Numpy array
-    :return: PIL instance (gray scale image in RGB space)
-    """
-    x = img.reshape(-1, 3)
-
-    M = (x - np.mean(x, axis=0))
-
-    latent, coeff = np.linalg.eig(np.cov(M.T))
-    res = np.dot(M, coeff[:, 0])
-
-    res = res[:, np.newaxis]
-    res = np.tile(res, (1, 1, 3))
-    res = np.reshape(res, img.shape)
-
-    res = image.array_to_img(res, scale=True)
-
-    return res
-
-
-def get_color_selectivity_index(filter, model, layer, idx_neuron, dataset_path):
+def get_color_selectivity_index(filter, model, layer, idx_neuron, dataset):
 
     activations = filter.get_activations()
     # filter.print_params()
@@ -73,11 +16,9 @@ def get_color_selectivity_index(filter, model, layer, idx_neuron, dataset_path):
     locations = filter.get_locations()
     max_rgb_activation = activations[0]
 
-
-
     if max_rgb_activation != 0.0:
 
-        images = load_images(dataset_path, image_names)
+        images = dataset.load_images(image_names)
         images_gray = []
 
         for i in xrange(len(images)):
@@ -85,7 +26,6 @@ def get_color_selectivity_index(filter, model, layer, idx_neuron, dataset_path):
             row_ini, row_fin, col_ini, col_fin = get_image_receptive_field(x, y, model, layer)
             init_image = images[i]
             img_crop = init_image[row_ini:row_fin+1, col_ini:col_fin+1]
-
 
             im_opp = rgb2opp(img_crop)
             im_gray = image2max_gray(im_opp)
@@ -100,10 +40,9 @@ def get_color_selectivity_index(filter, model, layer, idx_neuron, dataset_path):
         # new_activations.print_params()
         norm_gray_activations = new_activations/max_rgb_activation
 
-        print sum(norm_activations)
-        print sum(norm_gray_activations)
+        # print sum(norm_activations)
+        # print sum(norm_gray_activations)
         return 1 - (sum(norm_gray_activations)/sum(norm_activations))
-
     else:
         return None
