+++ nefesi/layer_data.py
@@ -2,7 +2,7 @@
 import numpy as np
 from nefesi.read_activations import get_sorted_activations, get_activations
 from nefesi.neuron_feature import compute_nf
-from nefesi.similarity_index import get_similarity_index, load_images
+from nefesi.similarity_index import get_similarity_index
 
 class LayerData(object):
 
@@ -23,26 +23,26 @@ class LayerData(object):
         self.filters = get_sorted_activations(file_names, images, model,
                                               self.layer_id, self.filters, num_max_activations, batch_size)
 
-    def build_neuron_feature(self, dataset_path, model):
-        self.filters = compute_nf(dataset_path, model, self.layer_id, self.filters)
+    def build_neuron_feature(self, dataset, model):
+        self.filters = compute_nf(dataset, model, self.layer_id, self.filters)
 
     def get_filters(self):
         return self.filters
 
-    def get_selectivity_idx(self, model, index_name, dataset_path, labels=None, **kwargs):
+    def get_selectivity_idx(self, model, index_name, dataset, labels=None, **kwargs):
         sel_idx = []
         for f in self.filters:
             if index_name == 'color':
-                res = f.color_selectivity_idx(model, self.layer_id, self.filters.index(f), dataset_path)
+                res = f.color_selectivity_idx(model, self.layer_id, self.filters.index(f), dataset)
                 sel_idx.append(res)
             elif index_name == 'orientation':
                 degrees = kwargs.get('degrees')
                 n_rotations = kwargs.get('n_rotations')
-                res = f.orientation_selectivity_idx(model, self.layer_id, self.filters.index(f), dataset_path,
+                res = f.orientation_selectivity_idx(model, self.layer_id, self.filters.index(f), dataset,
                                                     degrees, n_rotations)
                 sel_idx.append(res)
             elif index_name == 'symmetry':
-                res = f.symmetry_selectivity_idx(model, self.layer_id, self.filters.index(f), dataset_path)
+                res = f.symmetry_selectivity_idx(model, self.layer_id, self.filters.index(f), dataset)
                 sel_idx.append(res)
             elif index_name == 'class':
                 if labels is None:
@@ -52,7 +52,7 @@ class LayerData(object):
                 sel_idx.append(res)
         return sel_idx
 
-    def get_similarity_idx(self, model, dataset_path):
+    def get_similarity_idx(self, model, dataset):
         if self.similarity_index is not None:
             return self.similarity_index
         else:
@@ -66,7 +66,7 @@ class LayerData(object):
                 for i in xrange(size-1):
                     for j in xrange(i+1, size):
                         sim_idx = get_similarity_index(self.filters[i], self.filters[j], i, j,
-                                                       model, self.layer_id, dataset_path)
+                                                       model, self.layer_id, dataset)
                         self.similarity_index[i][j] = sim_idx
 
                 return self.similarity_index
@@ -75,7 +75,7 @@ class LayerData(object):
     def _decomposition_image(self, model, image):
 
         name_img = self.filters[0].get_images_id()[0]
-        image = load_images('/home/oprades/ImageNet/train/', [name_img])
+        # image = load_images('/home/oprades/ImageNet/train/', [name_img])
 
         activations = get_activations(model, image, print_shape_only=True, layer_name=self.layer_id)
